backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pika>=1.2.0
firebase-admin>=5.0.0
pydantic>=1.8.0
mysql-connector-python==8.0.28
